Fix input structure
