+++ src/heart/peripheral/sensor.py
@@ -7,7 +7,8 @@ from typing import Any, Iterator, NoReturn, Self
 import serial
 
 from heart.firmware_io import constants
-from heart.peripheral.core import Input, Peripheral
+from heart.peripheral import Input
+from heart.peripheral.core import Peripheral
 from heart.utilities.env import get_device_ports
 from heart.utilities.logging import get_logger
