+++ src/heart/peripheral/switch.py
@@ -7,8 +7,9 @@ from bleak.backends.device import BLEDevice
 
 from heart.firmware_io.constants import (BUTTON_LONG_PRESS, BUTTON_PRESS,
                                          SWITCH_ROTATION)
+from heart.peripheral import Input
 from heart.peripheral.bluetooth import UartListener
-from heart.peripheral.core import Input, Peripheral
+from heart.peripheral.core import Peripheral
 from heart.utilities.env import get_device_ports
 from heart.utilities.logging import get_logger
